Made scene serialise in reverse order to ensure correct order is maintained in file.
Worked on SceneManager including loading, and unloading of scenes, setting the current active scene, setting entry scene, etc.

Added file dialog utils to FileUtils

Got rid of APIENTRY macro redefinition issues with GLFW. Making sure to include windows if on windows in PCH before anything else is included.

Added context menu for scenes in hierarchy to Set Active Scene and Remove Scene from hierarchy.

Added open scene functionality for both normal and additive loading.

Reorganised project logic to be after Application initialised opposed to before.

Added GUI for SceneManager config so you can see all scenes in Project/Scenes folder and select which to include in project (to be used for final builds), and select from a combo out of the included scenes of which is the project's entry point.