Created Custom Asset File Watcher for Asset Directory.
This takes a snapshot every 5 seconds, and compares the delta to the prior snapshot to identify Add, Delete, Rename, Move, and Modify events for files in the Asset Directory for supported extension types.

When the event's are implemented it will be used to create, move, rename, and modify .meta files for assets in the AssetManager.

EFSW was not the right solution as it is quite janky regarding the inconsistency of event callbacks...